[Decomp][Script] Cohesion split: extract db script loaders to ScriptMgrDbScripts.cpp
